+++ CHANGELOG.md
@@ -33,7 +33,7 @@
 
 ### 🐛 Fixes
 - [207](https://github.com/vegaprotocol/data-node/pull/207) - Fix rewards schema and update vega dependencies to have reward event fixes
-
+- [239](https://github.com/vegaprotocol/data-node/pull/238) - Update GraphQL schema to not require every asset has a global reward account.
 
 ## 0.45.1
 *2021-10-23*